+++ kernel/cls/Cargo.toml
@@ -5,3 +5,4 @@ edition = "2021"
 
 [dependencies]
 cls_macros = { path = "cls_macros" }
+# preemption = { path = "../preemption" }
