+++ kernel/cls/cls_macros/src/lib.rs
@@ -1,13 +1,14 @@
 #![feature(proc_macro_diagnostic, proc_macro_span, let_chains)]
 
+mod int;
+mod user;
+
 use convert_case::{Case, Casing};
 use proc_macro::{Diagnostic, Level, Span, TokenStream};
-use quote::{quote, ToTokens};
+use quote::quote;
 use syn::{
     parse::{Parse, ParseStream},
-    parse_macro_input,
-    spanned::Spanned,
-    Attribute, Expr, Ident, LitInt, Token, Type, Visibility,
+    parse_macro_input, Attribute, Expr, Ident, LitInt, Token, Type, Visibility,
 };
 
 struct CpuLocal {
@@ -39,6 +40,15 @@ impl Parse for CpuLocal {
     }
 }
 
+/// A macro for declaring CPU-local variables.
+///
+/// Variables can either be an unsigned integer, or `cls::CpuLocalCell<T>`.
+/// Unsigned integers can be accessed without disabling preemption and should
+/// therefore be preferred.
+///
+/// In order to avoid circular dependencies, crates defining only integer
+/// CPU-locals can depend directly on `cls_macros`, avoiding `cls`'s dependency
+/// on `preemption`.
 #[proc_macro_attribute]
 pub fn cpu_local(args: TokenStream, input: TokenStream) -> TokenStream {
     // if !args.is_empty() {
@@ -79,22 +89,7 @@ pub fn cpu_local(args: TokenStream, input: TokenStream) -> TokenStream {
         Span::call_site().into(),
     );
 
-    let ((x64_asm_width, x64_reg_class), (aarch64_reg_modifier, aarch64_instr_width)) =
-        match ty.to_token_stream().to_string().as_ref() {
-            "u8" => (("byte", quote! { reg_byte }), (":w", "b")),
-            "u16" => (("word", quote! { reg }), (":w", "w")),
-            "u32" => (("dword", quote! { reg }), (":w", "")),
-            "u64" => (("qword", quote! { reg }), ("", "")),
-            _ => {
-                Diagnostic::spanned(ty.span().unwrap(), Level::Error, "invalid type")
-                    .help("CPU locals only support these types: `u8`, `u16`, `u32`, `u64`")
-                    .emit();
-                return TokenStream::new();
-            }
-        };
-
-    let x64_width_modifier = format!("{x64_asm_width} ptr ");
-    let x64_cls_location = format!("gs:[{offset}]");
+    let methods = int::methods(&ty, &offset).unwrap_or_else(|| user::methods(&ty, &offset));
 
     quote! {
         #(#attributes)*
@@ -110,150 +105,8 @@ pub fn cpu_local(args: TokenStream, input: TokenStream) -> TokenStream {
             _inner: ::core::cell::UnsafeCell<#ty>,
         }
 
-        #[cfg(target_arch = "x86_64")]
         impl #struct_name {
-            #[inline]
-            pub fn load(&self) -> #ty {
-                let ret;
-                unsafe {
-                    ::core::arch::asm!(
-                        ::core::concat!("mov {}, ", #x64_cls_location),
-                        out(#x64_reg_class) ret,
-                        options(preserves_flags, nostack),
-                    )
-                };
-                ret
-            }
-
-            #[inline]
-            pub fn increment(&self) {
-                unsafe {
-                    ::core::arch::asm!(
-                        ::core::concat!("inc ", #x64_width_modifier, #x64_cls_location),
-                        options(preserves_flags, nostack),
-                    )
-                };
-            }
-
-            #[inline]
-            pub fn decrement(&self) {
-                unsafe {
-                    ::core::arch::asm!(
-                        ::core::concat!("dec ", #x64_width_modifier, #x64_cls_location),
-                        options(preserves_flags, nostack),
-                    )
-                }
-            }
-        }
-
-        #[cfg(target_arch = "aarch64")]
-        impl #struct_name {
-            #[inline]
-            pub fn load(&self) -> #ty {
-                let ret;
-                unsafe {
-                    ::core::arch::asm!(
-                        "1:",
-                        // Load value.
-                        "mrs {tp_1}, tpidr_el1",
-                        concat!(
-                            "ldr", #aarch64_instr_width,
-                            " {ret", #aarch64_reg_modifier,"},",
-                            " [{tp_1},#", stringify!(#offset), "]",
-                        ),
-
-                        // Make sure task wasn't migrated between mrs and ldxr.
-                        "mrs {tp_2}, tpidr_el1",
-                        "cmp {tp_1}, {tp_2}",
-                        "b.ne 1b",
-
-                        tp_1 = out(reg) _,
-                        ret = out(reg) ret,
-                        tp_2 = out(reg) _,
-                    )
-                };
-                ret
-            }
-
-            #[inline]
-            pub fn add(&self, operand: #ty) {
-                unsafe {
-                    ::core::arch::asm!(
-                        "1:",
-                        // Load value.
-                        // TODO: Can we add offset and load in one instruction?
-                        "mrs {tp_1}, tpidr_el1",
-                        concat!("add {ptr}, {tp_1}, ", stringify!(#offset)),
-                        concat!("ldxr", #aarch64_instr_width, " {value", #aarch64_reg_modifier,"}, [{ptr}]"),
-
-                        // Make sure task wasn't migrated between msr and ldxr.
-                        "mrs {tp_2}, tpidr_el1",
-                        "cmp {tp_1}, {tp_2}",
-                        "b.ne 1b",
-
-                        // Compute and store value.
-                        "add {value}, {value}, {operand}",
-                        concat!("stxr", #aarch64_instr_width, " {cond:w}, {value", #aarch64_reg_modifier,"}, [{ptr}]"),
-
-                        // Make sure task wasn't migrated between ldxr and stxr.
-                        "cbnz {cond}, 1b",
-
-                        tp_1 = out(reg) _,
-                        ptr = out(reg) _,
-                        value = out(reg) _,
-                        tp_2 = out(reg) _,
-                        operand = in(reg) operand,
-                        cond = out(reg) _,
-
-                        options(nostack),
-                    );
-                }
-            }
-
-            pub fn sub(&self, operand: #ty) {
-                unsafe {
-                    ::core::arch::asm!(
-                        "1:",
-                        // Load value.
-                        // TODO: Can we add offset and load in one instruction?
-                        "mrs {tp_1}, tpidr_el1",
-                        concat!("add {ptr}, {tp_1}, ", stringify!(#offset)),
-                        concat!("ldxr", #aarch64_instr_width, " {value", #aarch64_reg_modifier,"}, [{ptr}]"),
-
-                        // Make sure task wasn't migrated between msr and ldxr.
-                        "mrs {tp_2}, tpidr_el1",
-                        "cmp {tp_1}, {tp_2}",
-                        "b.ne 1b",
-
-                        // Compute and store value.
-                        "sub {value}, {value}, {operand}",
-                        concat!("stxr", #aarch64_instr_width, " {cond:w}, {value", #aarch64_reg_modifier,"}, [{ptr}]"),
-
-                        // Make sure task wasn't migrated between ldxr and stxr.
-                        "cbnz {cond}, 1b",
-
-                        tp_1 = out(reg) _,
-                        ptr = out(reg) _,
-                        value = out(reg) _,
-                        tp_2 = out(reg) _,
-                        operand = in(reg) operand,
-                        cond = out(reg) _,
-
-                        options(nostack),
-                    );
-                }
-            }
-
-            #[inline]
-            pub fn increment(&self) {
-                self.add(1);
-            }
-
-            #[inline]
-            pub fn decrement(&self) {
-                self.sub(1);
-            }
-
+            #methods
         }
     }
     .into()
