+++ kernel/task/Cargo.toml
@@ -13,9 +13,9 @@ crossbeam-utils = { version = "0.8.2", default-features = false }
 
 irq_safety = { git = "https://github.com/theseus-os/irq_safety" }
 
+cls = { path = "../cls" }
 context_switch = { path = "../context_switch" }
 cpu = { path = "../cpu" }
-cpu_local = { path = "../cpu_local" }
 environment = { path = "../environment" }
 memory = { path = "../memory" }
 mod_mgmt = { path = "../mod_mgmt" }
